move sidebar to it's own feature folder
