+++ frontend/src/modules/navigation/Routes.tsx
@@ -16,10 +16,10 @@ import { useSelector } from 'react-redux';
 import { Switch } from 'react-router-dom';
 import { getParametersIsVentilating } from '../../store/controller/selectors';
 import AlarmLimitsPage from '../alarms/limits/AlarmLimitsPage';
-import LandingPageRoute from '../app/layouts/LandingPageRoute';
-import ScreensaverRoute from '../app/layouts/ScreensaverRoute';
-import SidebarRoute from '../app/layouts/SidebarRoute';
-import SidebarSlideRoute from '../app/layouts/SidebarSlideRoute';
+import LandingPageRoute from './layouts/LandingPageRoute';
+import ScreensaverRoute from './layouts/ScreensaverRoute';
+import SidebarRoute from './layouts/SidebarRoute';
+import SidebarSlideRoute from './layouts/SidebarSlideRoute';
 import DashboardPage from '../dashboard/DashboardPage';
 import LandingPage from '../landing-page/LandingPage';
 import LogsPage from '../logs/LogsPage';
