+++ frontend/src/modules/sidebar/index.ts
@@ -1,0 +1,2 @@
+export * from './Sidebar';
+export * from './SidebarClickable';
